requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
badges-io==1.0.8
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
discord.py==1.7.3
idna==3.2
multidict==5.1.0
Pillow==8.3.2
pip==21.2.4
psycopg2==2.9.1
requests==2.26.0
setuptools==56.0.0
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3